reflection-agent/pyproject.toml
[project]
name = "zenflection-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Mateusz Zbaraski",email = "notgivingyouthisone@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "langgraph (>=0.3.34,<0.4.0)",
    "grandalf (>=0.8,<0.9)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.14"

